[bug #161413610] separate models
